options now in a seperate ts file
